% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory setup}
\alias{directory setup}
\alias{setup_dir}
\title{Create and Fill a Forecasting Directory}
\usage{
setup_dir(
  main = ".",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information (and thus just the tidy messages).}
}
\value{
The \code{list} of directory settings \code{\link[base]{invisible}}-ly.
}
\description{
Combines \code{\link{create_dir}} and \code{\link{fill_dir}} to create a ready-to-run (via \code{\link{evercast}}) directory where indicated.
}
