% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory_settings}
\alias{directory_settings}
\title{Create a Directory Settings List}
\usage{
directory_settings(
  directory_config_file = "dir_config.yaml",
  subdirectories = c(forecasts = "forecasts", fits = "fits", models = "models",
    resources = "resources", data = "data"),
  save = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{directory_config_file}{\code{character} value of the path to the directory config YAML.}

\item{subdirectories}{\code{character} vector of the subdirectory names. Default includes \code{resources}, \code{data}, \code{models}, \code{fits}, and \code{forecasts}.}

\item{save}{\code{logical} indicator controlling if the output should be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing files should be updated (most users should leave as \code{TRUE}).}
}
\value{
Named \code{list} of settings for the directory.
}
\description{
Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for pipeline functionality.
}
