% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evercast.R
\docType{package}
\name{evercast}
\alias{evercast}
\title{Functions for Forecasting Everglades Wading Birds}
\description{
This package contains the functions used for forecasting everglades wading birds.
}
\keyword{package}
