% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prefabricated_count_datasets}
\alias{prefabricated_count_datasets}
\alias{prefab_count_datasets}
\alias{prefab_count_dataset_controls}
\title{Provide the Names or Controls for the Prefab Count Datasets}
\usage{
prefab_count_datasets()

prefab_count_dataset_controls()
}
\value{
\code{prefab_count_datasets}: \code{character} vector of count dataset names. \cr
        \code{prefab_count_dataset_controls}: \code{list} vector of count dataset controls. \cr
}
\description{
Create a \code{character} vector of the names of the pre-fabricated (prefab) count datasets or a \code{list} of their controls.
}
\examples{
 prefab_count_datasets()
 prefab_count_dataset_controls()

}
