% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prefabricated_covariate_datasets}
\alias{prefabricated_covariate_datasets}
\alias{prefab_covariate_datasets}
\alias{prefab_covariate_dataset_controls}
\title{Provide the Names or Controls for the Prefab Covariate Datasets}
\usage{
prefab_covariate_datasets()

prefab_covariate_dataset_controls()
}
\value{
\code{prefab_covariate_datasets}: \code{character} vector of covariatedataset names. \cr
        \code{prefab_covariate_dataset_controls}: \code{list} vector of covariate dataset controls. \cr
}
\description{
Create a \code{character} vector of the names of the pre-fabricated (prefab) covariate datasets or a \code{list} of their controls.
}
\examples{
 prefab_covariate_datasets()
 prefab_covariate_dataset_controls()

}
