% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prefabricated_datasets}
\alias{prefabricated_datasets}
\alias{prefab_datasets}
\alias{prefab_dataset_controls}
\title{Provide the Names or Controls for the Prefab Datasets}
\usage{
prefab_datasets()

prefab_dataset_controls()
}
\value{
\code{prefab_datasets}: \code{character} vector of dataset names. \cr
        \code{prefab_dataset_controls}: \code{list} vector of dataset controls. \cr
}
\description{
Create a \code{character} vector of the names of the pre-fabricated (prefab) datasets or a \code{list} of their controls.
}
\examples{
 prefab_datasets()
 prefab_dataset_controls()

}
