% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare covariate dataset}
\alias{prepare covariate dataset}
\alias{prepare_covariate_dataset}
\title{Prepare Covariate Data Tables for Forecasting}
\usage{
prepare_covariate_dataset(
  name = "all_total",
  main = ".",
  settings = directory_settings(),
  datafile = "Water/eden_covariates.csv",
  filename = "counts_all_total.csv",
  minyear = 1986,
  maxyear = this_year(),
  na.strings = "",
  regions = "all",
  save = TRUE,
  overwrite = TRUE,
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{name}{\code{character} name of the data set.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{datafile}{the path to the datafile within the folder for EvergladesWadingBird data}

\item{filename}{\code{character} value of the file for saving the output.}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{na.strings}{a character vector of strings which are to be
    interpreted as \code{\link{NA}} values.  Blank fields are also
    considered to be missing values in logical, integer, numeric and
    complex fields.  Note that the test happens \emph{after} 
    white space is stripped from the input, so \code{na.strings} 
    values may need their own white space stripped in advance.}

\item{regions}{\code{character} vector of the region(s) to include.}

\item{save}{\code{logical} indicator controlling if the output should be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing files should be updated (most users should leave as \code{TRUE}).}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be shown.}
}
\value{
\code{data.frame} for the specified data set.
}
\description{
Generate covariate datasets using existing functions.
}
