% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare covariates}
\alias{prepare covariates}
\alias{prepare_covariates}
\title{Prepare Covariate Data for the Evercast Repository}
\usage{
prepare_covariates(
  main = ".",
  covariate_datasets = prefab_covariate_datasets(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{covariate_datasets}{\code{character} vector of name(s) of covariate dataset(s) to include.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{verbose}{\code{logical} indicator of whether or not to print out all of the information or not (and thus just the tidy messages).}
}
\value{
\code{list} of prepared \code{datasets}.
}
\description{
Create specified \code{covariate_datasets} using their associated function and arguments.
}
