% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare max count dataset}
\alias{prepare max count dataset}
\alias{prepare_max_counts}
\title{Prepare Max Counts Data Tables for Forecasting}
\usage{
prepare_max_counts(
  name = "all_total",
  main = ".",
  settings = directory_settings(),
  species = "total",
  datafile = "Indicators/max_count_all_total.csv",
  filename = "counts_all_total.csv",
  minyear = 1986,
  maxyear = this_year(),
  level = "all",
  save = TRUE,
  overwrite = TRUE,
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{name}{\code{character} name of the data set.}

\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}

\item{filename}{\code{character} value of the file for saving the output.}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{level}{level at which to summarize (colony, subregion, region, all)}

\item{save}{\code{logical} indicator controlling if the output should be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing files should be updated (most users should leave as \code{TRUE}).}

\item{quiet}{\code{logical} indicator if progress messages should be quieted.}

\item{verbose}{\code{logical} indicator if detailed messages should be shown.}
}
\value{
\code{data.frame} for the specified data set.
}
\description{
Generate max count annual datasets using existing functions.
}
