% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{read and write count dataset controls}
\alias{read and write count dataset controls}
\alias{read_count_dataset_controls}
\alias{count_dataset_controls}
\alias{write_count_dataset_controls}
\title{Read and Write Dataset Count Control Lists}
\usage{
read_count_dataset_controls(main = ".", settings = directory_settings())

count_dataset_controls(
  main = ".",
  count_datasets = prefab_count_datasets(),
  settings = directory_settings()
)

write_count_dataset_controls(
  main = ".",
  new_count_dataset_controls = NULL,
  count_datasets = prefab_count_datasets(),
  settings = directory_settings(),
  quiet = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{count_datasets}{\code{character} vector of name(s) of rodent dataset(s) to include.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}

\item{new_dataset_controls}{\code{list} of controls for any new datasets (not in the prefab datasets) listed in \code{datasets} that are to be added to the control list and file.}
}
\value{
\code{list} of \code{datasets}' control \code{list}s, \code{\link[base]{invisible}}-ly for \code{write_dataset_controls}.
}
\description{
Input/Output functions for count dataset control lists.
}
