% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{read and write covariate dataset controls}
\alias{read and write covariate dataset controls}
\alias{read_covariate_dataset_controls}
\alias{covariate_dataset_controls}
\alias{write_covariate_dataset_controls}
\title{Read and Write Dataset Covariate Control Lists}
\usage{
read_covariate_dataset_controls(main = ".", settings = directory_settings())

covariate_dataset_controls(
  main = ".",
  covariate_datasets = prefab_covariate_datasets(),
  settings = directory_settings()
)

write_covariate_dataset_controls(
  main = ".",
  new_covariate_dataset_controls = NULL,
  covariate_datasets = prefab_covariate_datasets(),
  settings = directory_settings(),
  quiet = FALSE
)
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}}.}

\item{covariate_datasets}{\code{character} vector of name(s) of covariate dataset(s) to be created.}

\item{new_covariate_dataset_controls}{\code{list} of controls for any new datasets (not in the prefab datasets) listed in \code{datasets} that are to be added to the control list and file.}

\item{quiet}{\code{logical} indicator controlling if messages are printed.}
}
\value{
\code{list} of \code{datasets}' control \code{list}s, \code{\link[base]{invisible}}-ly for \code{write_dataset_controls}.
}
\description{
Input/Output functions for covariate dataset control lists.
}
