% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{read_data}
\alias{read_data}
\alias{read_counts}
\title{Read in and Format a Data File}
\usage{
read_data(
  main = ".",
  data_name = NULL,
  dataset = "all_total",
  datasets = prefab_count_datasets(),
  settings = directory_settings()
)

read_counts(main = ".", dataset = "all", settings = directory_settings())
}
\arguments{
\item{main}{\code{character} value of the name of the main component of the directory tree.}

\item{data_name}{\code{character} representation of the data needed. Current options include \code{"counts"}.}

\item{dataset, datasets}{\code{character} representation of the grouping name(s) used to define the data. \code{dataset} can only be length 1, \code{datasets} is not restricted in length.}

\item{settings}{\code{list} of controls for the directory, with defaults set in \code{\link{directory_settings}} that should generally not need to be altered.}
}
\value{
Data requested.
}
\description{
Read in a specified data file.
}
