% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{directory_settings}
\alias{directory_settings}
\title{Create a Directory Settings List}
\usage{
directory_settings(
  files = list(directory_config = "dir_config.yaml", count_dataset_controls =
    "count_dataset_controls.yaml", covariate_dataset_controls =
    "covariate_dataset_controls.yaml"),
  subdirectories = list(forecasts = "forecasts", fits = "fits", models = "models",
    resources = "resources", data = "data"),
  EvergladesWadingBird = list(source = "github", version = "latest"),
  save = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{files}{named \code{character} list of the file names (with extensions). Default includes \code{directory_config}, \code{count_dataset_controls}, \code{covariate_dataset_controls}.}

\item{subdirectories}{named \code{character} list of the subdirectory names. Default includes \code{resources}, \code{data}, \code{models}, \code{fits}, and \code{forecasts}.}

\item{EvergladesWadingBird}{\code{list} with \code{source} and \code{version} elements that are \code{character} values for the source and version of the Everglades Wading Bird to download. Default values retrieve the latest data from github. \cr \cr
See \code{\link[wader]{download_observations}}.}

\item{save}{\code{logical} indicator controlling if the output should be saved out.}

\item{overwrite}{\code{logical} indicator of whether or not the existing files should be updated (most users should leave as \code{TRUE}).}
}
\value{
Named \code{list} of settings for the directory.
}
\description{
Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for pipeline functionality.
}
